Vue 3 Datepicker - move out of src folder, tweak some CSS input defaults (#826)
