+++ src/components/KDateTimePicker/KDateTimePicker.spec.ts
@@ -3,7 +3,7 @@
 
 import { mount } from '@cypress/vue'
 import { format } from 'date-fns'
-import { TimePeriods, TimeframeKeys } from '@/utils/KDatePickerMockData'
+import { TimePeriods, TimeframeKeys } from '@mocks/KDateTimePickerMockData'
 import KDateTimePicker from '@/components/KDateTimePicker/KDateTimePicker.vue'
 
 const exampleTimeFrames = [
