+++ src/styles/_mixins.scss
@@ -18,3 +18,20 @@
 
   @return $string;
 }
+
+@mixin input-default {
+  background-color: var(--KInputBackground, var(--white, color(white)));
+  box-shadow: inset 0 0 0 1px var(--KInputBorder, var(--grey-300)) !important;
+  transition: color 0.1s ease, box-shadow 0.1s ease;
+}
+
+@mixin input-focus {
+  outline: none !important;
+  box-shadow: inset 0 0 0 1.5px var(--KInputFocus, var(--blue-400)) !important;
+  transition: all 0.1s ease;
+}
+
+@mixin input-hover {
+  box-shadow: inset 0 0 0 1px var(--KInputHover, var(--blue-200)) !important;
+  transition: all 0.1s ease;
+}
