+++ src/styles/forms/_inputs.scss
@@ -58,11 +58,9 @@
     line-height: 24px;
     border: none;
     border-radius: 3px;
-    background-color: var(--KInputBackground, var(--white, color(white)));
-    box-shadow: inset 0 0 0 1px var(--KInputBorder, var(--grey-300)) !important;
     box-sizing: border-box;
-    transition: color 0.1s ease, box-shadow 0.1s ease;
-
+    @include input-default;
+    
     &.k-input-small {
       font-size: var(--type-xs, type(xs));
       padding: var(--spacing-xs, spacing(xs)) var(--spacing-sm, spacing(sm));
@@ -74,8 +72,7 @@
     }
 
     &:hover {
-      box-shadow: inset 0 0 0 1px var(--KInputHover, var(--blue-200)) !important;
-      transition: all 0.1s ease;
+      @include input-hover;
 
       &.k-input-large {
         box-shadow: inset 0 0 0 2px var(--KInputHover, var(--blue-200)) !important;
@@ -84,9 +81,7 @@
     }
 
     &:focus {
-      outline: none !important;
-      box-shadow: inset 0 0 0 1.5px var(--KInputFocus, var(--blue-400)) !important;
-      transition: all 0.1s ease;
+      @include input-focus;
 
       &.k-input-large {
         box-shadow: inset 0 0 0 2px var(--KInputFocus, var(--blue-400)) !important;
